last commit with notification and alternate routes

+++ falconclient/app/src/main/java/ie/tcd/scss/ase/dataclasses/Wind.kt
@@ -1,7 +1,7 @@
 package ie.tcd.scss.ase.dataclasses
 
 data class Wind(
-	val degree: Int? = null,
+	val degree: Double? = null,
 	val speed: Double? = null,
 	val gust: Any? = null
 )
